Lint, move hook functions to separate file
