+++ src/components/scatterplot/ScatterplotSubscriber.js
@@ -7,7 +7,7 @@ import clamp from 'lodash/clamp';
 import TitleInfo from '../TitleInfo';
 import { VIEW_INFO } from '../../events';
 import { pluralize, capitalize } from '../../utils';
-import { useDeckCanvasSize, useReady, useUrls } from '../utils';
+import { useDeckCanvasSize, useReady, useUrls } from '../hooks';
 import { useCellsData, useCellSetsData, useExpressionMatrixData } from '../data-hooks';
 import { getCellColors } from '../interpolate-colors';
 import Scatterplot from './Scatterplot';
