+++ src/components/sets/CellSetSizesPlotSubscriber.js
@@ -2,7 +2,7 @@ import React, { useMemo, useEffect } from 'react';
 import TitleInfo from '../TitleInfo';
 import { useCoordination } from '../../app/state/hooks';
 import { COMPONENT_COORDINATION_TYPES } from '../../app/state/coordination';
-import { useUrls, useReady, useGridItemSize } from '../utils';
+import { useUrls, useReady, useGridItemSize } from '../hooks';
 import { useCellSetsData } from '../data-hooks';
 import { treeToSetSizesBySetNames } from './reducer';
 import CellSetSizesPlot from './CellSetSizesPlot';
